requirements.txt
requests>=2.31.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
colorama>=0.4.6
beautifulsoup4>=4.12.0
loguru>=0.7.0
pyyaml>=6.0.0
PyJWT>=2.8.0
python-socketio[client]>=5.11.0
